requirements.txt
discord
twitch-python
